requirements.txt
torch>=2.1.0
transformers>=4.51.0
accelerate>=0.26.0
faiss-cpu>=1.7.4
datasets>=2.18.0
numpy>=1.26.0
tqdm>=4.66.0
